Extracting a module for the Utility methods
I was encountering a situation where the module was being defined
elsewhere, which meant I didn't have the methods.
