+++ spec/factories/create_curation_concern_factory.rb
@@ -2,7 +2,7 @@
 #   pid = CurationConcern.mint_a_pid
 #   curation_concern = factory_name.to_s.classify.constantize.new(pid: pid)
 #   attributes = override_attributes.reverse_merge(FactoryGirl.attributes_for(factory_name))
-#   actor = CurationConcern.actor(curation_concern, user, attributes)
+#   actor = CurationConcern::Utility.actor(curation_concern, user, attributes)
 #   actor.create!
 #   curation_concern
 # end
